style: move testutils into a separate module
