+++ cmd/check_test.go
@@ -5,8 +5,8 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/everdrone/grab/internal/testutils"
 	"github.com/everdrone/grab/internal/utils"
+	"github.com/everdrone/grab/testutils"
 )
 
 const fileOk = `
