+++ cmd/find_test.go
@@ -6,8 +6,8 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/everdrone/grab/internal/testutils"
 	"github.com/everdrone/grab/internal/utils"
+	"github.com/everdrone/grab/testutils"
 )
 
 func TestFindCmd(t *testing.T) {
