+++ internal/instance/cache_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 
 	"github.com/everdrone/grab/internal/config"
-	"github.com/everdrone/grab/internal/testutils"
+	"github.com/everdrone/grab/testutils"
 )
 
 func TestBuildSiteCache(t *testing.T) {
