+++ internal/instance/download_test.go
@@ -7,8 +7,8 @@ import (
 	"testing"
 
 	"github.com/everdrone/grab/internal/config"
-	"github.com/everdrone/grab/internal/testutils"
 	"github.com/everdrone/grab/internal/utils"
+	"github.com/everdrone/grab/testutils"
 )
 
 func TestDownload(t *testing.T) {
